renamed package names, applied String.format suggestion
